+++ cmd/resources/resource_cmds.go
@@ -1,32 +1,31 @@
 // this file WILL be generated (sc-241153)
 
-package cmd
+package resources
 
 import (
-	"net/http"
-
 	"github.com/spf13/cobra"
 
-	"ldcli/cmd/resources"
 	"ldcli/internal/analytics"
+	"ldcli/internal/resources"
 )
 
-func addAllResourceCmds(rootCmd *cobra.Command, client *http.Client, analyticsTracker analytics.Tracker) {
+func AddAllResourceCmds(rootCmd *cobra.Command, client resources.Client, analyticsTracker analytics.Tracker) {
 	// Resource commands
-	gen_TeamsResourceCmd := resources.NewResourceCmd(
+	gen_TeamsResourceCmd := NewResourceCmd(
 		rootCmd,
 		analyticsTracker,
 		"teams",
-		"A team is a group of members in your LaunchDarkly account.",
-		"A team can have maintainers who are able to add and remove team members. It also can have custom roles assigned to it that allows shared access to those roles for all team members. To learn more, read [Teams](https://docs.launchdarkly.com/home/teams).\n\nThe Teams API allows you to create, read, update, and delete a team.\n\nSeveral of the endpoints in the Teams API require one or more member IDs. The member ID is returned as part of the [List account members](/tag/Account-members#operation/getMembers) response. It is the `_id` field of each element in the `items` array.",
+		"Make requests (list, create, etc.) on teams",
+		"A team is a group of members in your LaunchDarkly account. A team can have maintainers who are able to add and remove team members. It also can have custom roles assigned to it that allows shared access to those roles for all team members. To learn more, read [Teams](https://docs.launchdarkly.com/home/teams).\n\nThe Teams API allows you to create, read, update, and delete a team.\n\nSeveral of the endpoints in the Teams API require one or more member IDs. The member ID is returned as part of the [List account members](/tag/Account-members#operation/getMembers) response. It is the `_id` field of each element in the `items` array.",
 	)
 
 	// Operation commands
-	resources.NewOperationCmd(gen_TeamsResourceCmd, client, resources.OperationData{
+	NewOperationCmd(gen_TeamsResourceCmd, client, OperationData{
 		Short: "Create team",
 		Long:  "Create a team. To learn more, read [Creating a team](https://docs.launchdarkly.com/home/teams/creating).\n\n### Expanding the teams response\nLaunchDarkly supports four fields for expanding the \"Create team\" response. By default, these fields are **not** included in the response.\n\nTo expand the response, append the `expand` query parameter and add a comma-separated list with any of the following fields:\n\n* `members` includes the total count of members that belong to the team.\n* `roles` includes a paginated list of the custom roles that you have assigned to the team.\n* `projects` includes a paginated list of the projects that the team has any write access to.\n* `maintainers` includes a paginated list of the maintainers that you have assigned to the team.\n\nFor example, `expand=members,roles` includes the `members` and `roles` fields in the response.\n",
 		Use:   "create", // TODO: translate post -> create
-		Params: []resources.Param{
+
+		Params: []Param{
 			{
 				Name:        "expand",
 				In:          "query",
@@ -38,12 +37,11 @@ func addAllResourceCmds(rootCmd *cobra.Command, client *http.Client, analyticsTr
 		RequiresBody: true,
 		Path:         "/api/v2/teams",
 	})
-
-	resources.NewOperationCmd(gen_TeamsResourceCmd, client, resources.OperationData{
+	NewOperationCmd(gen_TeamsResourceCmd, client, OperationData{
 		Short: "Get team",
 		Long:  "Fetch a team by key.\n\n### Expanding the teams response\nLaunchDarkly supports four fields for expanding the \"Get team\" response. By default, these fields are **not** included in the response.\n\nTo expand the response, append the `expand` query parameter and add a comma-separated list with any of the following fields:\n\n* `members` includes the total count of members that belong to the team.\n* `roles` includes a paginated list of the custom roles that you have assigned to the team.\n* `projects` includes a paginated list of the projects that the team has any write access to.\n* `maintainers` includes a paginated list of the maintainers that you have assigned to the team.\n\nFor example, `expand=members,roles` includes the `members` and `roles` fields in the response.\n",
 		Use:   "get",
-		Params: []resources.Param{
+		Params: []Param{
 			{
 				Name:        "teamKey", // TODO: kebab case/trim key? to be consistent with our existing flags (e.g. projectKey = project)
 				In:          "path",