+++ internal/analytics/client.go
@@ -30,6 +30,20 @@ type Tracker interface {
 		optOut bool,
 		step string,
 	)
+	SendSetupSDKSelectedEvent(
+		accessToken,
+		baseURI string,
+		optOut bool,
+		sdk string,
+	)
+	SendSetupFlagToggledEvent(
+		accessToken,
+		baseURI string,
+		optOut,
+		on bool,
+		count int,
+		duration_ms int64,
+	)
 }
 
 type Client struct {
@@ -151,6 +165,44 @@ func (c *Client) SendSetupStepStartedEvent(
 	)
 }
 
+func (c *Client) SendSetupSDKSelectedEvent(
+	accessToken,
+	baseURI string,
+	optOut bool,
+	sdk string,
+) {
+	c.sendEvent(
+		accessToken,
+		baseURI,
+		optOut,
+		"CLI Setup SDK Selected",
+		map[string]interface{}{
+			"sdk": sdk,
+		},
+	)
+}
+
+func (c *Client) SendSetupFlagToggledEvent(
+	accessToken,
+	baseURI string,
+	optOut,
+	on bool,
+	count int,
+	duration_ms int64,
+) {
+	c.sendEvent(
+		accessToken,
+		baseURI,
+		optOut,
+		"CLI Setup Flag Toggled",
+		map[string]interface{}{
+			"on":          on,
+			"count":       count,
+			"duration_ms": duration_ms,
+		},
+	)
+}
+
 func (a *Client) Wait() {
 	a.wg.Wait()
 }
@@ -178,6 +230,24 @@ func (c *NoopClient) SendSetupStepStartedEvent(
 	baseURI string,
 	optOut bool,
 	step string,
+) {
+}
+
+func (c *NoopClient) SendSetupSDKSelectedEvent(
+	accessToken,
+	baseURI string,
+	optOut bool,
+	sdk string,
+) {
+}
+
+func (c *NoopClient) SendSetupFlagToggledEvent(
+	accessToken,
+	baseURI string,
+	optOut,
+	on bool,
+	count int,
+	duration_ms int64,
 ) {
 }
 
@@ -242,6 +312,44 @@ func (m *MockTracker) SendSetupStepStartedEvent(
 		"CLI Setup Step Started",
 		map[string]interface{}{
 			"step": step,
+		},
+	)
+}
+
+func (m *MockTracker) SendSetupSDKSelectedEvent(
+	accessToken,
+	baseURI string,
+	optOut bool,
+	sdk string,
+) {
+	m.sendEvent(
+		accessToken,
+		baseURI,
+		optOut,
+		"CLI Setup SDK Selected",
+		map[string]interface{}{
+			"sdk": sdk,
+		},
+	)
+}
+
+func (m *MockTracker) SendSetupFlagToggledEvent(
+	accessToken,
+	baseURI string,
+	optOut,
+	on bool,
+	count int,
+	duration_ms int64,
+) {
+	m.sendEvent(
+		accessToken,
+		baseURI,
+		optOut,
+		"CLI Setup Flag Toggled",
+		map[string]interface{}{
+			"on":          on,
+			"count":       count,
+			"duration_ms": duration_ms,
 		},
 	)
 }